+++ L2_GameServer_c5/java/net/sf/l2j/gameserver/handler/itemhandlers/FishItem.java
@@ -17,11 +17,11 @@ import net.sf.l2j.gameserver.serverpackets.SystemMessage;
  */
 public class FishItem implements IItemHandler
 {
-	private static short[] _itemIds = null;
+	private static int[] _itemIds = null;
 	public FishItem()
 	{
 		FishTable ft = FishTable.getInstance();
-        _itemIds = new short[ft.GetFishItemCount()];
+        _itemIds = new int[ft.GetFishItemCount()];
         for (int i = 0; i < ft.GetFishItemCount(); i++)
         {
             _itemIds[i] = ft.getFishIdfromList(i);
@@ -55,7 +55,7 @@ public class FishItem implements IItemHandler
 			activeChar.sendMessage("Fish failed to open!");
 		}
 	}
-	public void giveItems(L2PcInstance activeChar, short itemId, int count)
+	public void giveItems(L2PcInstance activeChar, int itemId, int count)
     {
         activeChar.addItem("FishItem", itemId, count, null, false);
         
@@ -77,7 +77,7 @@ public class FishItem implements IItemHandler
     {
 		player.destroyItemByItemId("FishItem", itemId, 1, null, false);
     }
-	public short[] getItemIds()
+	public int[] getItemIds()
 	{
 		return _itemIds;
 	}
