+++ L2_GameServer_c5/java/net/sf/l2j/gameserver/handler/itemhandlers/FishShots.java
@@ -19,8 +19,8 @@ import net.sf.l2j.gameserver.util.Broadcast;
 public class FishShots implements IItemHandler 
 { 
 	// All the item IDs that this handler knows.
-	private static short[] _itemIds = { 6535, 6536, 6537, 6538, 6539, 6540 }; 
-	private static short[] _skillIds = { 2181, 2182, 2183, 2184, 2185, 2186 };
+	private static int[] _itemIds = { 6535, 6536, 6537, 6538, 6539, 6540 }; 
+	private static int[] _skillIds = { 2181, 2182, 2183, 2184, 2185, 2186 };
 
 	/* (non-Javadoc) 
  	* @see net.sf.l2j.gameserver.handler.IItemHandler#useItem(net.sf.l2j.gameserver.model.L2PcInstance, net.sf.l2j.gameserver.model.L2ItemInstance) 
@@ -78,7 +78,7 @@ public class FishShots implements IItemHandler
         activeChar.setTarget(oldTarget);        
 	} 
 
-	public short[] getItemIds() 
+	public int[] getItemIds() 
 	{ 
 		return _itemIds; 
 	}
