+++ L2_GameServer_c5/java/net/sf/l2j/gameserver/handler/itemhandlers/Remedy.java
@@ -36,7 +36,7 @@ import net.sf.l2j.gameserver.serverpackets.SystemMessage;
 
 public class Remedy implements IItemHandler
 {
-	private static short[] _itemIds = { 1831, 1832, 1833, 1834, 3889 };
+	private static int[] _itemIds = { 1831, 1832, 1833, 1834, 3889 };
 	
 	public void useItem(L2PlayableInstance playable, L2ItemInstance item)
 	{
@@ -134,7 +134,7 @@ public class Remedy implements IItemHandler
             playable.destroyItem("Consume", item.getObjectId(), 1, null, false);
        }
 	}
-	public short[] getItemIds()
+	public int[] getItemIds()
 	{
 		return _itemIds;
 	}
